+++ feature/library/src/main/java/com/github/gericass/world_heritage_client/library/PlaylistId.kt
@@ -1,6 +1,6 @@
 package com.github.gericass.world_heritage_client.library
 
-enum class PlaylistId(val id: Long) {
+enum class PlaylistId(val id: Int) {
     HISTORY(-1),
     FAVORITE(-2),
     LATER(-3)
